bin/requirements.txt
pyqt5
pyqt5-tools
python-docx